+++ clients/vscode/src/extension.ts
@@ -5,7 +5,7 @@
 'use strict';
 
 import * as path from 'path';
-import { ExtensionContext, commands, workspace } from 'vscode';
+import { ExtensionContext, commands } from 'vscode';
 
 import {
   LanguageClient,
@@ -71,12 +71,12 @@ export function activate(context: ExtensionContext) {
   const serverOptions: ServerOptions = {
     run: {
       module: serverModule,
-      transport: TransportKind.ipc,
+      transport: TransportKind.stdio,
       options: runOptions,
     },
     debug: {
       module: serverModule,
-      transport: TransportKind.ipc,
+      transport: TransportKind.stdio,
       options: debugOptions,
     },
   };
@@ -85,15 +85,6 @@ export function activate(context: ExtensionContext) {
   const clientOptions: LanguageClientOptions = {
     // Register the server for kos documents
     documentSelector: [{ scheme: 'file', language: 'kos' }],
-
-    synchronize: {
-      fileEvents: workspace.createFileSystemWatcher(
-        '**/*.ks',
-        true,
-        true,
-        false,
-      ),
-    },
 
     errorHandler: {
       error(error: Error, message: Message, count: number): ErrorAction {